server/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
fastapi==0.88.0
h11==0.14.0
idna==3.4
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python-decouple==3.6
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
